+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 * removed plugin export-image from core
 * instanciate all plugins first, then run their init() functions
 * Important: Please run scripts/migrate/1.8.0.sql and composer install after updating.
+* Chart editor design update
 
 ### 1.7.11
 * fixed symlink creation in plugin installation